fix the rest of dup issue
