+++ crates/doc/src/parser/item.rs
@@ -1,4 +1,4 @@
-use crate::{Comments, solang_ext::SafeUnwrap};
+use crate::{Comments, helpers::function_signature, solang_ext::SafeUnwrap};
 use solang_parser::pt::{
     ContractDefinition, ContractTy, EnumDefinition, ErrorDefinition, EventDefinition,
     FunctionDefinition, StructDefinition, TypeDefinition, VariableDefinition,
@@ -172,22 +172,7 @@ impl ParseSource {
     /// Get the signature of the source (for functions, includes parameter types)
     pub fn signature(&self) -> String {
         match self {
-            Self::Function(func) => {
-                let name = func.name.as_ref().map_or(func.ty.to_string(), |n| n.name.to_owned());
-                if !func.params.is_empty() {
-                    format!(
-                        "{}({})",
-                        name,
-                        func.params
-                            .iter()
-                            .map(|p| p.1.as_ref().map(|p| p.ty.to_string()).unwrap_or_default())
-                            .collect::<Vec<_>>()
-                            .join(",")
-                    )
-                } else {
-                    name
-                }
-            }
+            Self::Function(func) => function_signature(func),
             _ => self.ident(),
         }
     }
